Use the packet trace timestamps in netflow records
Instead of the current time.
Equally, copy the incoming stream (on the in channel) timestamps.

The aggregate segment will take the first timestamp as flow start and
the last timestamp as flow end from all the connected sub-flows for the
final flow (at least when the packet trace/flow data is sorted by time).