CI: Merge the container image CIs into one workflow
To be able to introduce a dependency between the base flowpipeline and
visualization builds. This is required to ensure that the visualization
container uses the correct flowpipeline version on a specified tag.

And dependencies between workflows are not implemented in Forgejo
Actions yet [1], so they have to be jobs in the same workflow.

[1] https://codeberg.org/forgejo/forgejo/issues/6494