CI: Avoid duplicate Renovate execution on PR merge
Triggering on closed PRs and on pushes to main creates two runs at the
same time. This commits avoids that by not triggering on pushes to main
and instead by relying on the closed PR trigger. This way, Renovate can
quickly add a comment to closed (not merged) PRs and also rebase open
Renovate PRs against latest main.
Manual pushes to main subvert this, but they are discouraged anyways.