ci: use pre-built container image for commitlint
Although the upstream documentation [1] recommends the previous setup,
we currently run into too many rate limits by npmjs. This may be because
we don't have functioning caching and are not running on GitHub (which
probably has higher rate limits, as they are owned by the same company).

Using the container image is our poor man's way to cache the commitlint
executable between runs.

[1] https://commitlint.js.org/guides/ci-setup.html