ci: fix container image builds on 2+ labels
With the switch from using a dedicated action to just using the docker
container in 81a6bad, this workflow failed on releases. The cause is
that only then, container images get more than one label (v*.*.* and
latest), and those labels are separated with newline instead of spaces
or tabs.

The change makes the tag processing special-character-proof, as it reads
per line, not accepting any field separators inside of lines. And
container images are only allowed to have alphanumeric and some common
extra characters, and that is enforced by the docker metadata action.